Snapshot vote power tests (#1749)
* initial building of the tests and fixing of the cannonfile

* implement majority of tests for SnapshotRecord

* fix bugs in the recent tests

* most changes in

* fix lint issues

---------

+++ protocol/governance/contracts/interfaces/external/ISnapshotRecord.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.0;
 
 interface ISnapshotRecord {
-    function balanceOfOnPeriod(address account, uint periodId) external view returns (uint);
+    function balanceOfOnPeriod(address account, uint128 periodId) external view returns (uint);
 
-    function totalSupplyOnPeriod(uint periodId) external view returns (uint);
+    function totalSupplyOnPeriod(uint128 periodId) external view returns (uint);
 
     function takeSnapshot(uint128 id) external;
 }
