+++ protocol/governance/contracts/storage/SnapshotVotePowerEpoch.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 library SnapshotVotePowerEpoch {
     struct Data {
-        uint256 snapshotId;
+        uint128 snapshotId;
         mapping(address => uint256) recordedVotingPower;
     }
 }
